create workaround route to trigger Initial Data Insertion (scmboy)
